+++ question3/src/com/san/carModels/Sedan/Sedan.java
@@ -9,7 +9,7 @@ public class Sedan extends Car {
         String brand, 
         String color, 
         int year, 
-        float mileage,
+        double mileage,
         double trunkSize
     ) {
         super (name, brand,color, year, mileage, trunkSize);
